ANTLR4 to validate spec, provide grammar to validate access expressions
This change adds ANTLR4 to the build for two uses. First, I copied in
the ANTLR4 ABNF grammar file[1] and used that to parse the ABNF
specification that we created to validate that the specification is
syntactically correct. Second, I created an ANTLR4 grammar file for
access expressions and used that in a test to validate that the
grammar is correct. The ANTLR4 grammar file (AccessExpression.g4)
can be used by other projects that are written in a language supported
by ANTLR4 (c++, c#, python, etc.) to validate that the access expressions
that they are creating are valid.

[1] https://github.com/antlr/grammars-v4/blob/master/abnf/Abnf.g4