Add ANTLR grammar and tests to contrib directory
Added contrib/antlr4 directory which contains an ANTLRv4 grammar
file for parsing AccessExpressions. I also pulled in the ANTLRv4 ABNF
grammar file[1] and used that to parse the ABNF specification that we
created to validate that the specification is syntactically correct. I created
tests to validate that the AccessExpression grammar is correct. The
ANTLRv4 grammar file (AccessExpression.g4) can be used by other
projects that are written in a language supported by ANTLR4 (c++, c#,
python, etc.) to validate that the access expressions that they are
creating are valid.

[1] https://github.com/antlr/grammars-v4/blob/master/abnf/Abnf.g4