Avoids looping when evaluating single set of authorizations
Simplified the code to avoid looping when there is a single
set of authorizations.  This change was made in an attempt to
improve performance and it did very slightly.

Before this change saw the following.

```
Benchmark                                     Mode  Cnt   Score   Error   Units
AccessExpressionBenchmark.measureEvaluation  thrpt   12  13.609 ± 0.300  ops/us
```

And after this change saw the following.

```
Benchmark                                     Mode  Cnt   Score   Error   Units
AccessExpressionBenchmark.measureEvaluation  thrpt   12  13.806 ± 0.254  ops/us
```

The performance improvement is slight so maybe not worthwhile.  However
this change simplified the evaluator code a good bit, so that is
worthwhile.