memoizes creating a parse tree
This change offers a new parse() method on AccessExpression that creates
a parse tree if one does not exits.  This change allows decoupling how
access expression are created (w/ or w/o an initial parse tree) from
using parse trees.

For example code the like the following that wanted to use a parse tree
to validate an access expression would have required the
ParsedAccessExpression type to be passed to it.  Now it can accept an
AccessExpression and call parse() which will do the most efficient thing
depending on how the expression was created.

```java

void checkExpression(AccessExpression expression){
    var parsed = expression.parse();
    // use parse tree to validate expression
}

```