Added AccessEvaluator.canAccess(ParsedAccessExpression)
In the MultiAccessEvaluatorImpl case it was calling
AccessEvaluator.canAccess(byte[]) for each AccessEvaluator.
canAccess(byte[]) would re-parse the AccessExpression each
time. This change enables the MultiAccessEvaluatorImpl
to get the ParsedAccessExpression and call the new
method on each AccessEvaluator. The new method traverses
the parse tree for evalation so that parsing is not
performed multiple times. The downside is that there are
two different methods in use, but the tests were already
checking this and still pass.

Closes #78